Various fixes for lambda's #17
There are a bunch of problems for lambda's which I've been working on.  At some
point, I wonder whether there is a better solution.  The essential problems are:

(1) we need to enforce the return type constraints of indirect invocations
through f_apply(), etc.

(2) we need to assert the return type constraints for lambda accesses to ensure
they then flow as expected.

For functions, I have done this now and it seems to work.  However, for methods,
it's more complicated as these are currently implememted using procedure calls,
rather than via the function interface.