Implement default framing! #48
This implements default framing for reasoning about methods.  Given the current
syntax of Whiley, this is the best we can do.  Specifically, any reference
passed into a method is assumed to be modified by the method.  All other
references are assumed unchanged.  Likewise, any reference returned by the
method was either in the original frame, or was allocated new.