Improvements to Framing #60
This provides a straightforward (and sometimes frustating) solution for loop
framing.  Essentially, any reference which is accessed within a loop is havoced.
The key challenge is to recognise that only variables read or written in a loop
could affect its frame.

In addition, have significantlly updated the manner in which framing is handled.
Instead of using a custom #frame method for every user-defined type, we now have
a single generic method Any#within which does the same thing.  One useful
difference is this allows us to talk about the relationship of references in the
heap without knowing exactly what their type is.

At this stage, there remain outstanding problems with loop framing.  In
particular, variables which are used as "roots" for framing need to be extracted
into loop invariant temporaries which are assigned before the loop.  Otherwise,
as their value gets updated within the loop, this then devalues the loop
invariant and results in the postcondition being unprovable.