Fix HEAP as global variable
This adjusts the way that the HEAP is encoded into Boogie.  Instead of
chaing a HEAP variable through procedures, it is now a single global
variable with procedures updated with "modifies HEAP".

There are some benefits to this approach.  Firstly, the generated Boogie
is actually clearer.  Secondly, we can leverage Boogie's existing old()
syntax for refering to the heap.  Furthermore, there don't appear to be
any downsides that have been uncovered.