Redesign encoding of functions
This completely redesigns the way that functions are encoded into
Boogie.  Instead of always encoding them as Boogie functions they are
now encoded as both a Boogie function and a procedure.  More
specifically, all invocations to a function in a non-specification
element are extracted as procedure calls.  However, in specification
elements, they are left as function calls (because there is not choice).
A connection is established between the function and procedure view of a
Whiley function.