Implement Short Circuit Semantics
This implements short circuit semantics for expressions involving function or
method calls.  This unfortunately requires extracting expressions into chains of
if statements.  It's ugly ... but it works.

In the end, the handling of loops requires the use of unstructured control flow
in the case of a loop condition involving a function call.  In fact, it will
default to unstructured control flow by default at this stage.  A flag is
available for enabling the use of structured control flow in places where
possible.  Furthermore, the translation strategy is a little conservative as
there are cases where structured control flow could be used, but are not.