Implement static variable assignments 138
This puts through a minimal implementation for static variable assignments.
Since Whiley has no current support for modifies clauses, it simply assume that
*any* static variable can be modified by a method.

There are few interesting questions as to which statc variables should be
included in the default any clause.  For example, private statics in external
modules are currently not included.