Problem running through tests
I have encountered a problem running through all of the tests.  There seem to be
a small number of tests which consume a lot of resources and eventually take
down eclipse.  Potentially this arises because these tests are no longer valid,
but its hard to say.  The problem seems to be that when they timeout, they don't
additionally kill their child z3 process as well.

One approach to managing these difficult cases would be to have a full abort
option.